refactor util functions into logical groups
