+++ package.json
@@ -25,7 +25,7 @@
 	],
 	"keywords": [
 		"tmc",
-		"test", 
+		"test",
 		"my",
 		"code"
 	],
