+++ src/ui/templateEngine.ts
@@ -5,7 +5,7 @@ import * as vscode from "vscode";
 
 import { SubmissionResultReport, TmcLangsTestResult } from "../api/types";
 import Resources from "../config/resources";
-import { getProgressBar, numbersToString } from "../utils";
+import { getProgressBar, numbersToString } from "../utils/utils";
 
 export default class TemplateEngine {
     private cssPath: string;
